# docs/swagger.yaml
definitions:
...
  clientError.ErrorResponse:
    properties:
      error:
        type: string
      error_code:
        type: string
    type: object
...
  model.CreateMoniestRequest:
    properties:
      bio:
        description: optional
        type: string
      card_id:
        type: string
      description:
        description: optional
        type: string
      fee:
        type: number
      message:
        description: optional
        type: string
    required:
    - card_id
    - fee
    type: object
...
  model.Moniest:
    properties:
      bio:
        type: string
      description:
        type: string
      id:
        type: string
      score:
        type: number
      subscription_info:
        $ref: '#/definitions/model.SubscriptionInfo'
    type: object
  model.OwnUser:
    properties:
      background_photo_link:
        type: string
      background_photo_thumbnail_link:
        type: string
      created_at:
        type: string
      email:
        type: string
      email_verified:
        type: boolean
      id:
        type: string
      location:
        type: string
      moniest:
        $ref: '#/definitions/model.Moniest'
      name:
        type: string
      profile_photo_link:
        type: string
      profile_photo_thumbnail_link:
        type: string
      surname:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
...
  model.SubscriptionInfo:
    properties:
      fee:
        type: number
      message:
        type: string
      updated_at:
        type: string
    type: object
  model.UpdateMoniestProfileRequest:
    properties:
      bio:
        type: string
      description:
        type: string
      subscription_info:
        $ref: '#/definitions/model.SubscriptionInfo'
    type: object
...
info:
  contact: {}
paths:
...
  /moniests:
    post:
      consumes:
      - application/json
      description: Turn into moniest
      parameters:
      - description: ' '
        in: body
        name: CreateMoniestBody
        required: true
        schema:
          $ref: '#/definitions/model.CreateMoniestRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.OwnUser'
        "400":
          description: user is already moniest
          schema:
            $ref: '#/definitions/clientError.ErrorResponse'
        "403":
          description: 'forbidden operation: email is not verified'
          schema:
            $ref: '#/definitions/clientError.ErrorResponse'
        "404":
          description: not found user
          schema:
            $ref: '#/definitions/clientError.ErrorResponse'
        "406":
          description: invalid body
          schema:
            $ref: '#/definitions/clientError.ErrorResponse'
        "500":
          description: server error
          schema:
            $ref: '#/definitions/clientError.ErrorResponse'
      security:
      - bearerAuth: []
      summary: Be Moniest
      tags:
      - Moniest
  /moniests/:username/subscribe:
    post:
      consumes:
      - application/json
      description: Subscribe to Moniest
      parameters:
      - description: moniest username
        in: path
        name: username
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: already subscribed
          schema:
            $ref: '#/definitions/clientError.ErrorResponse'
        "403":
          description: subscribe own
          schema:
            $ref: '#/definitions/clientError.ErrorResponse'
        "404":
          description: moniest is not found
          schema:
            $ref: '#/definitions/clientError.ErrorResponse'
        "500":
          description: server error
          schema:
            $ref: '#/definitions/clientError.ErrorResponse'
      security:
      - bearerAuth: []
      summary: Subscribe to Moniest
      tags:
      - Moniest
  /moniests/:username/unsubscribe:
    post:
      consumes:
      - application/json
      description: Unsubscribe from Moniest
      parameters:
      - description: moniest username
        in: path
        name: username
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: user not subscribed
          schema:
            $ref: '#/definitions/clientError.ErrorResponse'
        "403":
          description: unsubscribe own
          schema:
            $ref: '#/definitions/clientError.ErrorResponse'
        "404":
          description: moniest is not found
          schema:
            $ref: '#/definitions/clientError.ErrorResponse'
        "500":
          description: server error
          schema:
            $ref: '#/definitions/clientError.ErrorResponse'
      security:
      - bearerAuth: []
      summary: Unsubscribe from Moniest
      tags:
      - Moniest
...
  /moniests/profile:
    patch:
      consumes:
      - application/json
      description: Update Moniest Profile details
      parameters:
      - description: all fields are optional
        in: body
        name: UpdateMoniestBody
        required: true
        schema:
          $ref: '#/definitions/model.UpdateMoniestProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.OwnUser'
        "403":
          description: user is not moniest
          schema:
            $ref: '#/definitions/clientError.ErrorResponse'
        "404":
          description: user is not found
          schema:
            $ref: '#/definitions/clientError.ErrorResponse'
        "406":
          description: invalid body | invalid bio | invalid desc | invalid fee | invalid
            message
          schema:
            $ref: '#/definitions/clientError.ErrorResponse'
        "500":
          description: server error
          schema:
            $ref: '#/definitions/clientError.ErrorResponse'
      security:
      - bearerAuth: []
      summary: Update Moniest Profile
      tags:
      - Moniest
...
securityDefinitions:
  bearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"